moved spectrometer forward model to dedicate diagnostics module
