+++ docs/source/models.rst
@@ -2,5 +2,7 @@ Edge profile models
 ===================
 
 .. autoclass:: pedinf.models.mtanh
+   :members: prediction, gradient, jacobian
 
 .. autoclass:: pedinf.models.lpm
+   :members: prediction, gradient, jacobian, prediction_and_jacobian
